requirements.txt
fastapi
uvicorn
kafka-python
faker
psycopg2-binary
email-validator